Camunda implementation
The Camunda now does the work instead of telling it to do something. The file `camunda.js` is the worker that automatically checks whether there are any tasks available to it if there are then this worker decides what to do with each task. For example TicketCreation process now is handled by Camunda and everything goes through Camunda. Camunda decides who made the ticket and then the ticket is sent to the appropirate place.